+++ src/app/(public)/mate/_components/map/mapDetail.tsx
